chore: fix features path and rename jam api lib file
